Rename assets, restructure drawing, add spritesheets
Renamed opponent_worm to enemy_worm, as well as changed the _x postfix to
_dead for the player and the worms.
Also changed the way entities are drawn in the level and added a class
to deal with spritesheets.
Since there are no sprite sheets in the repository to use for testing, this class is
untested and thus likely to contain bugs.